chore: refactor the status/condition updating into an interface with functions (#73)
This change begins refactoring the reconciler code. The existing
reconciler code makes heavy use of embedded structs to mimic an
inheritance model. Go doesn't have an inheritance model, and generally
discourages this in favor of interfaces.

This code moves the `updateStatus`, `updateCondition`, `setReadyStatus`
and `refetch` methods out of the `BaseReconciler{}` struct and into
their own packages where they are pure functions. These functions work
by checking that the objects passed in match the `hasStatusReconciler`
interface.

The ultimate goal is to create generic `RequestReconciler` and
`TemplateReconciler` controllers that can be configured dynamically on
startup to decide which custom resources they are monitoring for, and
which `IBuilder` is used for them. This pattern will make it simpler to
add in our own custom controllers in the future, or even support a
pluggable controller model.

_Note_: The naming of `hasStatusReconciler` is ugly - but it's a private
interface so we can refactor it any time once we have a better name.