chore: new request/template specific status interfaces, and a new package for setting conditions (#75)
This is mostly code-cleanup in prep for a larger change coming. We
create new `IRequestStatus` and `ITemplateStatus` interfaces and move
some specific methods into them. We've also moved all of the
`UpdateCondition()` calls to the `internal/controllers/internal/status`
package - so that the logic is all confined in one place for what types
of strings and values are passed into the conditions.