feat: refactor to create generic RequestReconciler and new IBuilder interface (#77)
The main goal for this PR is to begin the process of simplifying the
reconciler logic and standardizing it enough that we can reuse a single
reconciler for all Access Requests. The motivation is to stop
duplicating the Reconciler logic (and tests), and instead make it easier
to build new Access Request and Access Template resource types.

* The original Builders structs have been moved into a `legacybuilders`
package
* A new `builders` package has been created that defines an `IBuilder`
interface with a few required methods: `GetTemplate()`,
`GetAccessDuration()`, `SetOwnerReference()`, `CreateAccessResources()`
and `AccessResourcesAreReady()`.
* A new `builders/utils` package has been created with common functions
that most or all of the builders will ultimately use
* A new `builders/execaccessbuilder` package implements an IBuilder for
the `ExecAccessRequest` CR
* A new `controllers/requestcontroller` package has been created that
creates a new generic `RequestReconciler{}` struct that can be
configured to watch any CRD and use an `IBuilder` interface to execute
the resource build.
* New `controllers/internal/status` and
`controllers/internal/ctrlrequeue` packages have been created to make
resource status updates and requeueing less error prone.